pyproject.toml
[project]
name = "twitter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.13.0",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "click==8.3.2",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "fastapi-cli==0.0.24",
    "fastapi-cloud-cli==0.16.1",
    "fastapi[standard]==0.135.3",
    "fastar==0.10.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "jinja2==3.1.6",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "psycopg2-binary==2.9.11",
    "pwdlib==0.3.0",
    "pycparser==3.0",
    "pydantic==2.13.0",
    "pydantic-core==2.46.0",
    "pydantic-extra-types==2.11.2",
    "pydantic-settings==2.13.1",
    "pygments==2.20.0",
    "pyjwt==2.12.1",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.26",
    "pyyaml==6.0.3",
    "rich==15.0.0",
    "rich-toolkit==0.19.7",
    "rignore==0.7.6",
    "ruff==0.15.10",
    "sentry-sdk==2.57.0",
    "shellingham==1.5.4",
    "sqlalchemy==2.0.49",
    "sqlmodel==0.0.38",
    "starlette==1.0.0",
    "typer==0.24.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uvicorn==0.44.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websockets==16.0",
]
